fixed deleted user cookie

+++ src/routes/(user)/leaderboard/+page.server.ts
@@ -47,17 +47,20 @@ export async function load(load_event: ServerLoadEvent): Promise<any> {
   if (res) {
     if (
       res.rowCount == 0 ||
-      (res?.rowCount !== 0 && is_object_empty(res.rows[0]) !== false)
+      (res?.rowCount !== 0 && (is_object_empty(res.rows[0]) !== false || res.rows[0]["uid"] === null))
     ) {
       get(other_error_logger_store).error(
         "\nError parsing db function result at (user)/leaderboard/+page.server.ts:52.\n" +
-          res
+        res
       );
-      return error(500);
+
+      delete_user_cookie(load_event.cookies);
+
+      return redirect(303, "/login");
     }
+
     data.details = res.rows[0];
   } else {
-    delete_user_cookie(load_event.cookies);
     return error(500);
   }
 
@@ -85,7 +88,7 @@ export async function load(load_event: ServerLoadEvent): Promise<any> {
     if (res.rowCount !== 0 && is_object_empty(res.rows[0]) !== false) {
       get(other_error_logger_store).error(
         "\nError parsing db function result at (user)/leaderboard/+page.server.ts:85.\n" +
-          res
+        res
       );
       return error(500);
     }
@@ -103,7 +106,7 @@ export async function load(load_event: ServerLoadEvent): Promise<any> {
     if (res2.rowCount !== 0 && is_object_empty(res2.rows[0]) !== false) {
       get(other_error_logger_store).error(
         "\nError parsing db function result at (user)/leaderboard/+page.server.ts:103.\n" +
-          res2
+        res2
       );
       return error(500);
     }
